server/requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
Django==5.2.4
django-cors-headers==4.7.0
django-elasticsearch-dsl==8.2
django-filter==25.1
django-nine==0.2.7
djangorestframework==3.16.0
drf-nested-routers==0.94.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
elastic-transport==8.17.1
elasticsearch==8.18.1
elasticsearch-dsl==8.18.0
factory_boy==3.3.3
Faker==37.4.2
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
